+++ packages/dialect-react-ui/components/common/index.ts
@@ -1,0 +1,3 @@
+export * from './primitives';
+
+export { default as NetworkBadge } from './NetworkBadge';