+++ packages/dialect-react/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## [UNRELEASED]
+
+- Add encrypted threads support with Sollet wallet
+
 ## [0.2.0] - 2022-03-31
 
 - Add email notifications support